+++ src/main.rs
@@ -10,13 +10,10 @@ mod config;
 mod error;
 mod git;
 mod plan;
-mod publish;
 mod push;
 mod release;
 mod replace;
 mod shell;
-mod tag;
-mod util;
 mod version;
 
 fn main() {
