--- src/util.rs
@@ -1,8 +1,0 @@
-pub fn resolve_bool_arg(yes: bool, no: bool) -> Option<bool> {
-    match (yes, no) {
-        (true, false) => Some(true),
-        (false, true) => Some(false),
-        (false, false) => None,
-        (_, _) => unreachable!("clap should make this impossible"),
-    }
-}
